# docs/swagger-old.yaml
openapi: 3.0.0
info:
  title: APIServiceGo
  version: 1.0.0
  description: EV Charging API for searching charging connectors, getting estimates, managing orders, and processing payments for electric vehicle charging sessions.
servers:
  - url: http://{{localhost}}
paths:
  /v1/search:
    post:
      operationId: searchChargingConnectors
      summary: Search Charging Connectors
      tags:
        - Discovery
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Page size for pagination.
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              example:
                total: 1
                page: 1
                per_page: 100
                catalogs:
                  - id: catalog-ev-charging-001
                    provider:
                      id: ecopower-charging
                      descriptor:
                        name: EcoPower Charging Pvt Ltd
                      address:
                        name: MG JVLR Jogeshwari Caves Road
                        geo_coordinates: [12.9716, 77.5946]
                    rating:
                      value: 4.5
                      count: 128
                    connectors:
                      - id: ev-charger-ccs2-001
                        isActive: true
                        availabilityWindow:
                          - startTime: "06:00:00"
                            endTime: "22:00:00"
                        price:
                          value: "8"
                          currency: "INR/kWH"
                        connectorAttributes:
                          connectorType: TYPE_2
                          maxPowerKW: 60
                          minPowerKW: 5
                          socketCount: 2
                          reservationSupported: true
                          status: Available
                          chargingSpeed: FAST
                          powerType: DC
                          connectorFormat: CABLE
                    offers:
                      - id: offer-ccs2-60kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - CCS2 60kW
                        items:
                          - ev-charger-ccs2-001
                        price:
                          currency: INR
                          value: 18
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-01T00:00:00Z"
                          endDate: "2026-03-31T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 2.5
                          idleFeePolicy: ₹2/min after 10 min post-charge
                        provider: ecopower-charging
                      - id: offer-ccs2-120kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - CCS2 120kW
                        items:
                          - ev-charger-ccs2-002
                        price:
                          currency: INR
                          value: 22
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-15T00:00:00Z"
                          endDate: "2026-04-15T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                          - NetBanking
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 2
                          idleFeePolicy: ₹3/min after 15 min post-charge
                        provider: greencharge-koramangala
                      - id: offer-type2-22kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - Type 2 22kW
                        items:
                          - ev-charger-type2-001
                        price:
                          currency: INR
                          value: 15
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-20T00:00:00Z"
                          endDate: "2026-04-20T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 1.5
                          idleFeePolicy: ₹1/min after 30 min post-charge
                        provider: powergrid-indiranagar
        "400":
          description: Bad Request - Invalid request parameters or malformed request body.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid request parameters or malformed request body."
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "422":
          description: Unprocessable Entity - Validation error in request data.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation error in request data."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: >-
        Search for available charging connectors and offers near a location or
        by filters. Either evse_id or geo_coordinates with distance_meters is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
...
  /v1/orders/{order_id}/payment:
    post:
      operationId: initiatePayment
      summary: Initiate Payment
      tags:
        - Payments
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Bpp-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
              example:
                order:
                  id: order-bpp-789012
                  status: ACTIVE
                  mode: RESERVATION
                amount:
                  value: 128.64
                  currency: INR
                beneficiaryId: ""
                acceptedPaymentMethod:
                  - BankTransfer
                  - UPI
                  - Wallet
                paymentUrl: ""
                validity:
                  startDate: "2025-01-27T00:00:00Z"
                  endDate: "2025-04-27T23:59:59Z"
        "400":
          description: Bad Request - Invalid request parameters or malformed request body.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid request parameters or malformed request body."
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - Order not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Order not found."
        "409":
          description: Conflict - Payment already processed or order in invalid state for payment.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "CONFLICT"
                  message: "Payment already processed or order in invalid state for payment."
        "422":
          description: Unprocessable Entity - Validation error in payment data.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation error in payment data."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: Initiate or fetch payment instructions for the specified order.
      requestBody:
        content:
          application/json:
            schema:
              type: object
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          default: <string>
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_REQUEST"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
    AvailabilityWindow:
      type: object
      description: Time window when the connector is available
      properties:
        startTime:
          type: string
          format: time
          description: Start time in HH:mm:ss format
        endTime:
          type: string
          format: time
          description: End time in HH:mm:ss format
    Rating:
      type: object
      description: User rating information
      properties:
        value:
          type: number
          format: float
          description: Average rating value
        count:
          type: integer
          description: Number of ratings
    Address:
      type: object
      description: Address information
      properties:
        name:
          type: string
          description: Address name/location
        geo_coordinates:
          type: array
          minItems: 2
          maxItems: 2
          description: Geographic coordinates as [latitude, longitude] array
          items:
            type: number
          example: [12.9716, 77.5946]
    ProviderDescriptor:
      type: object
      description: Provider descriptor
      properties:
        name:
          type: string
          description: Provider name
    Provider:
      type: object
      description: Provider information
      properties:
        id:
          type: string
          description: Provider identifier
        descriptor:
          $ref: "#/components/schemas/ProviderDescriptor"
        address:
          $ref: "#/components/schemas/Address"
    ConnectorAttributes:
      type: object
      description: Technical attributes of the connector
      properties:
        connectorType:
          type: string
          enum:
            - TYPE_1
            - TYPE_2
          description: Type of connector
        maxPowerKW:
          type: number
          format: float
          description: Maximum power in kilowatts
        minPowerKW:
          type: number
          format: float
          description: Minimum power in kilowatts
        socketCount:
          type: integer
          description: Number of sockets available
        reservationSupported:
          type: boolean
          description: Whether reservations are supported
        ocppId:
          type: string
          description: OCPP identifier
        evseId:
          type: string
          description: EVSE identifier
        parkingType:
          type: string
          description: Type of parking location
        connectorId:
          type: string
          description: Connector identifier
        powerType:
          type: string
          enum:
            - DC
            - AC
          description: Power type
        connectorFormat:
          type: string
          enum:
            - CABLE
            - OTHERS
          description: Connector format
        chargingSpeed:
          type: string
          enum:
            - SLOW
            - NORMAL
            - FAST
          description: Charging speed category
        status:
          type: string
          enum:
            - Available
            - Occupied
            - Reserved
            - OutOfOrder
            - Unknown
          description: Current status of the connector
        amenityFeature:
          type: array
          description: List of available amenities
          items:
            type: string
        roamingNetwork:
          type: string
          description: Roaming network identifier
    Connector:
      type: object
      description: Charging connector information
      properties:
        id:
          type: string
          description: Unique connector identifier
        isActive:
          type: boolean
          description: Whether the connector is currently active
        availabilityWindow:
          type: array
          description: Time windows when the connector is available
          items:
            $ref: "#/components/schemas/AvailabilityWindow"
        connectorAttributes:
          $ref: "#/components/schemas/ConnectorAttributes"
    BuyerFinderFee:
      type: object
      description: Finder fee information
      properties:
        feeType:
          type: string
          description: Type of fee (e.g., PERCENTAGE)
        feeValue:
          type: number
          format: float
          description: Fee value
    OfferAttributes:
      type: object
      description: Additional offer attributes
      properties:
        buyerFinderFee:
          $ref: "#/components/schemas/BuyerFinderFee"
        idleFeePolicy:
          type: string
          description: Idle fee policy description
    ApplicableQuantity:
      type: object
      description: Quantity unit for pricing
      properties:
        unitText:
          type: string
          description: Human-readable unit name
        unitCode:
          type: string
          description: Unit code
        unitQuantity:
          type: number
          format: float
          description: Quantity value
    Price:
      type: object
      description: Pricing information
      properties:
        currency:
          type: string
          description: Currency code (ISO 4217)
        value:
          type: number
          format: float
          description: Price value
        applicableQuantity:
          $ref: "#/components/schemas/ApplicableQuantity"
    Validity:
      type: object
      description: Offer validity period
      properties:
        startDate:
          type: string
          format: date-time
          description: Start date and time
        endDate:
          type: string
          format: date-time
          description: End date and time
    OfferDescriptor:
      type: object
      description: Offer description
      properties:
        name:
          type: string
          description: Offer name
    Offer:
      type: object
      description: Pricing offer for connectors
      properties:
        id:
          type: string
          description: Unique offer identifier
        descriptor:
          $ref: "#/components/schemas/OfferDescriptor"
        items:
          type: array
          description: List of connector IDs this offer applies to
          items:
            type: string
        price:
          $ref: "#/components/schemas/Price"
        validity:
          $ref: "#/components/schemas/Validity"
        acceptedPaymentMethod:
          type: array
          description: List of accepted payment methods
          items:
            type: string
        offerAttributes:
          $ref: "#/components/schemas/OfferAttributes"
        provider:
          type: string
          description: Provider identifier
    Catalog:
      type: object
      description: Catalog containing connectors and offers
      properties:
        id:
          type: string
          description: Unique catalog identifier
        provider:
          $ref: "#/components/schemas/Provider"
        rating:
          $ref: "#/components/schemas/Rating"
        connectors:
          type: array
          description: List of charging connectors
          items:
            $ref: "#/components/schemas/Connector"
        offers:
          type: array
          description: List of pricing offers for connectors
          items:
            $ref: "#/components/schemas/Offer"
    SearchResponse:
      type: object
      required:
        - total
        - page
        - per_page
        - catalogs
      properties:
        total:
          type: integer
          description: Total number of catalogs matching the search criteria
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Number of items per page
        catalogs:
          type: array
          description: List of catalogs containing connectors and offers
          items:
            $ref: "#/components/schemas/Catalog"
    Vehicle:
      type: object
      description: Vehicle information
      properties:
        make:
          type: string
          description: Vehicle manufacturer
        model:
          type: string
          description: Vehicle model
        type:
          type: string
          enum:
            - 2-wheeler
            - 3-wheeler
            - 4-wheeler
          description: Vehicle type
    SearchFilters:
      type: object
      description: Search filters
      properties:
        cpo:
          type: string
          description: Charge Point Operator identifier
        connector_type:
          type: string
          enum:
            - TYPE_1
            - TYPE_2
          description: Type of connector
        max_power_kw:
          type: number
          format: float
          minimum: 0
          description: Minimum maximum power in kilowatts (filters connectors with maxPowerKW >= this value)
        amenities:
          type: array
          description: List of required amenities (filters connectors that have all specified amenities)
          items:
            type: string
          example:
            - Restaurant
            - Restroom
            - Wi-Fi
        vehicle:
          $ref: "#/components/schemas/Vehicle"
    SearchSort:
      type: object
      description: Sorting criteria
      properties:
        sort_by:
          type: string
          description: Field to sort by
        order:
          type: string
          enum:
            - asc
    TimeWindow:
      type: object
      description: Desired time window
      properties:
        start:
          type: string
          format: date-time
          description: Start date and time
        end:
          type: string
          format: date-time
          description: End date and time
    SearchRequest:
      type: object
      properties:
        evse_id:
          type: string
          description: Specific EVSE identifier to look up
        geo_coordinates:
          type: array
          minItems: 2
          maxItems: 2
          description: Geographic coordinates as [latitude, longitude] array
          items:
            type: number
          example: [12.9716, 77.5946]
        distance_meters:
          type: number
          format: float
          minimum: 0
          maximum: 50000
          example: 5000
          description: Radius of search in meters (0-50000)
        time_window:
          $ref: "#/components/schemas/TimeWindow"
        filters:
          $ref: "#/components/schemas/SearchFilters"
        sort:
          $ref: "#/components/schemas/SearchSort"
      oneOf:
        - type: object
          required:
            - evse_id
          properties:
            evse_id:
              type: string
              description: Specific EVSE identifier to look up
            time_window:
              $ref: "#/components/schemas/TimeWindow"
            filters:
              $ref: "#/components/schemas/SearchFilters"
        - type: object
          required:
            - geo_coordinates
            - distance_meters
          properties:
            geo_coordinates:
              type: array
              minItems: 2
              maxItems: 2
              description: Geographic coordinates as [latitude, longitude] array
              items:
                type: number
              example: [12.9716, 77.5946]
            distance_meters:
              type: number
              format: float
              minimum: 0
              maximum: 50000
              example: 5000
              description: Radius of search in meters (0-50000)
            time_window:
              $ref: "#/components/schemas/TimeWindow"
            filters:
              $ref: "#/components/schemas/SearchFilters"
...
    Amount:
      type: object
      description: Amount for estimation
      properties:
        value:
          type: number
          format: float
          minimum: 0
          description: Amount value
        currency:
          type: string
          description: Currency code (ISO 4217)
...
    OrderInfo:
      type: object
      description: Order information
      properties:
        id:
          type: string
          description: Order identifier
        mode:
          type: string
          description: Order mode (e.g., reservation, RESERVATION)
        status:
          type: string
          description: Order status (e.g., quoted_price, ACTIVE, CANCELLED, COMPLETED)
...
    PaymentResponse:
      type: object
      description: Payment response
      properties:
        order:
          $ref: "#/components/schemas/OrderInfo"
        amount:
          $ref: "#/components/schemas/Amount"
          description: Total payment amount
        beneficiaryId:
          type: string
          description: Beneficiary identifier
        acceptedPaymentMethod:
          type: array
          description: List of accepted payment methods
          items:
            type: string
        paymentUrl:
          type: string
          format: uri
          description: URL for payment gateway (if applicable)
        validity:
          $ref: "#/components/schemas/Validity"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []
...